Change fetch args to more compact syntax for normal flows
Originally I was trying to balance between a native JS experience and the shape of AWS SecretsManager. However this is
hitting a point where that is no longer feasible with the caching client being built in a separate PR. This commit
changes the shape to match that upcoming shape.

As a side benefit the syntax really cleans up nicely for the normal call pattern of only needing a secretId